external file for editable key navugation type, use link on function without lambda, ignore file on serialization
