+++ components/keyedit/impl/src/main/java/com/flipperdevices/keyedit/impl/api/KeyEditFeatureEntryImpl.kt
@@ -1,7 +1,6 @@
 package com.flipperdevices.keyedit.impl.api
 
 import android.net.Uri
-import android.os.Bundle
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.navigation.NavGraphBuilder
@@ -11,16 +10,15 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
 import com.flipperdevices.core.di.AppGraph
-import com.flipperdevices.core.ktx.android.parcelable
 import com.flipperdevices.core.ui.navigation.ComposableFeatureEntry
 import com.flipperdevices.keyedit.api.KeyEditFeatureEntry
 import com.flipperdevices.keyedit.api.NotSavedFlipperKey
 import com.flipperdevices.keyedit.impl.composable.ComposableEditScreen
 import com.flipperdevices.keyedit.impl.model.EditableKey
+import com.flipperdevices.keyedit.impl.model.navigation.EditableKeyType
 import com.flipperdevices.keyedit.impl.viewmodel.KeyEditViewModel
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
-import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import tangle.viewmodel.compose.tangleViewModel
@@ -78,27 +76,11 @@ class KeyEditFeatureEntryImpl @Inject constructor() : KeyEditFeatureEntry {
                 viewModel,
                 title = title,
                 state = state,
-                onBack = {
-                    navController.popBackStack()
-                },
+                onBack = navController::popBackStack,
                 onSave = {
-                    viewModel.onSave { navController.popBackStack() }
+                    viewModel.onSave(navController::popBackStack)
                 }
             )
         }
-    }
-}
-
-class EditableKeyType : NavType<EditableKey>(isNullableAllowed = false) {
-    override fun get(bundle: Bundle, key: String): EditableKey? {
-        return bundle.parcelable(key)
-    }
-
-    override fun parseValue(value: String): EditableKey {
-        return Json.decodeFromString(value)
-    }
-
-    override fun put(bundle: Bundle, key: String, value: EditableKey) {
-        bundle.putParcelable(key, value)
     }
 }
